- Added shadow influence logic.
Presentation tip: usign of combined task at shadowsview.cpp.
Presentation tip: this looks like it is needed to invent an 'inside-out'
design.

+++ AmberUI/shadowstructure.h
@@ -58,7 +58,7 @@ struct WorldPosition
 typedef std::map<Element::ElementType, int> ShadowStructure;
 typedef std::function<ShadowStructure(ShadowStructure, Direction::DirectionType)> ShadowVariator;
 
-typedef int ShadowInfluence;
+typedef double ShadowInfluence;
 
 struct Shadow
 {
